Custom Qualities

+++ app/lib/cron/cronRenamer.py
@@ -1,12 +1,11 @@
-from app.config.db import Movie, History, RenameHistory, Session as Db
+from app.config.db import Movie, RenameHistory, Session as Db, MovieQueue
 from app.lib.cron.cronBase import cronBase
-from stat import ST_MODE
+from app.lib.qualities import Qualities
 import fnmatch
 import logging
 import os
 import re
 import shutil
-import stat
 import time
 
 log = logging.getLogger(__name__)
@@ -23,6 +22,11 @@ class RenamerCron(cronBase):
     minimalFileSize = 1024 * 1024 * 10 # 10MB
     ignoredInPath = ['_unpack', '.appledouble', '/._'] #unpacking, smb-crap
 
+    # Filetypes
+    movieExt = ['*.mkv', '*.wmv', '*.avi', '*.mpg', '*.mpeg', '*.mp4', '*.m2ts', '*.iso']
+    nfoExt = ['*.nfo']
+    subExt = ['*.sub', '*.srt', '*.idx', '*.ssa', '*.ass']
+
     def conf(self, option):
         return self.config.get('Renamer', option)
 
@@ -34,7 +38,7 @@ class RenamerCron(cronBase):
         if not os.path.isdir(self.conf('download')):
             log.info("Watched folder doesn't exist.")
 
-        time.sleep(10)
+        #time.sleep(10)
         while True and not self.abort:
             now = time.time()
             if (self.lastChecked + self.intervalSec) < now:
@@ -43,7 +47,7 @@ class RenamerCron(cronBase):
                 self.doRename()
                 self.running = False
 
-            time.sleep(1)
+            time.sleep(5)
 
         log.info('Renamer has shutdown.')
 
@@ -63,6 +67,9 @@ class RenamerCron(cronBase):
 
         allFiles = self.findFiles()
 
+        if allFiles:
+            log.debug("Ready to rename some files.")
+
         for files in allFiles:
 
             # See if imdb link is in nfo file
@@ -78,13 +85,24 @@ class RenamerCron(cronBase):
                 movie = self.determineMovie(files)
 
             if movie:
-                finalDestination = self.renameFiles(files, movie)
+                finalDestination = self.renameFiles(files, movie['movie'], movie['queue'])
                 if self.config.get('Renamer', 'trailerQuality').lower() != 'false':
-                    self.trailerQueue.put({'id': movie.imdb, 'movie': movie, 'destination':finalDestination})
+                    self.trailerQueue.put({'id': movie['movie'].imdb, 'movie': movie['movie'], 'destination':finalDestination})
             else:
                 log.info('No Match found for: %s' % str(files['files']))
 
-    def renameFiles(self, files, movie):
+    def getQueue(self, movie):
+
+        log.info('Finding quality for %s.' % movie.name)
+
+        # Assuming quality is the top most, as that should be the last downloaded..
+        for queue in movie.queue:
+            if(queue.name):
+                return queue
+
+        return False
+
+    def renameFiles(self, files, movie, queue = None):
         '''
         rename files based on movie data & conf
         '''
@@ -105,6 +123,13 @@ class RenamerCron(cronBase):
         if moviename[:3].lower() == 'the':
             namethe = moviename[3:] + ', The'
 
+        #quality
+        if not queue:
+            queue = self.getQueue(movie)
+        quality = Qualities.types[queue.qualityType]['label']
+        if not quality:
+            quality = ''
+
         replacements = {
              'cd': '',
              'cdNr': '',
@@ -112,10 +137,23 @@ class RenamerCron(cronBase):
              'namethe': namethe.strip(),
              'thename': moviename.strip(),
              'year': movie.year,
-             'first': namethe[0].upper()
+             'first': namethe[0].upper(),
+             'quality': quality,
         }
         if multiple:
             cd = 1
+
+        justAdded = []
+
+        totalSize = 0
+        for file in files['files']:
+            fullPath = os.path.join(file['path'], file['filename'])
+            totalSize += os.path.getsize(fullPath)
+
+            # Do something with ISO, as they should be between DVDRip and BRRIP
+            ext = os.path.splitext(file['filename'])[1].lower()[1:]
+            if ext == 'iso':
+                totalSize -= (os.path.getsize(fullPath) / 1.6)
 
         finalDestination = None
         for file in files['files']:
@@ -132,26 +170,30 @@ class RenamerCron(cronBase):
             old = os.path.join(file['path'], file['filename'])
             dest = os.path.join(destination, folder, filename)
 
-            log.info('Moving file "%s" to %s.' % (old, dest))
-            
             finalDestination = os.path.dirname(dest)
             if not os.path.isdir(finalDestination):
-                mode = os.stat(destination)
- 
+
                 # Use same permissions as conf('destination') folder
                 try:
-                    chmod = stat.S_IMODE(mode[ST_MODE])
-                    os.makedirs(finalDestination, chmod) #chmod doesn't work properly here?
-                    os.chmod(finalDestination, chmod)
-                    log.error('Creating directory %s, chmod: %d' % (finalDestination, chmod))
-                except OSError:
+                    #mode = os.stat(destination)
+                    #chmod = mode[ST_MODE] & 07777
+                    log.info('Creating directory %s' % finalDestination)
                     os.makedirs(finalDestination)
+                    shutil.copymode(destination, finalDestination)
+                    #os.chmod(finalDestination, chmod)
+                except OSError:
                     log.error('Failed setting permissions for %s' % finalDestination)
+                    os.makedirs(finalDestination)
 
-            if not os.path.isfile(dest):
+            # Remove old if better quality
+            removed = self.removeOld(os.path.join(destination, folder), justAdded, totalSize)
+
+            if not os.path.isfile(dest) and removed:
+                log.info('Moving file "%s" to %s.' % (old, dest))
                 shutil.move(old, dest)
+                justAdded.append(dest)
             else:
-                log.error('File %s already exists.' % filename)
+                log.error('File %s already exists or not better.' % filename)
                 break
 
             #get subtitle if any & move
@@ -173,9 +215,47 @@ class RenamerCron(cronBase):
                 cd += 1
 
         # Mark movie downloaded
-        movie.status = u'downloaded'
+        if queue.markComplete:
+            movie.status = u'downloaded'
+
+        queue.completed = True
 
         return finalDestination
+
+    def removeOld(self, path, dontDelete = [], newSize = 0):
+
+        files = []
+        oldSize = 0
+        for root, subfiles, filenames in os.walk(path):
+
+            for filename in filenames:
+                ext = os.path.splitext(filename)[1].lower()[1:]
+                fullPath = os.path.join(root, filename)
+                oldSize += os.path.getsize(fullPath)
+
+                # iso's are huge, but the same size as 720p, so remove some filesize for better comparison
+                if ext == 'iso':
+                    oldSize -= (os.path.getsize(fullPath) / 1.6)
+
+                if not fullPath in dontDelete:
+
+                    # Only delete media files
+                    if '*.' + ext in self.movieExt and not '-trailer' in filename:
+                        files.append(fullPath)
+
+        log.info('Quality Old: %s, New %s.' % (int(oldSize / 1024 / 1024), int(newSize / 1024 / 1024)))
+        if oldSize < newSize:
+            for file in files:
+                try:
+                    os.remove(file)
+                    log.info('Removed old file: %s' % file)
+                except OSError:
+                    log.info('Couldn\'t delete file: %s' % file)
+            return True
+        else:
+            log.info('New file(s) are smaller then old ones, don\'t overwrite')
+            return False
+
 
     def doReplace(self, string, replacements):
         '''
@@ -205,7 +285,7 @@ class RenamerCron(cronBase):
 
     def determineMovie(self, files):
         '''
-        Try find movie based on folder names and History table
+        Try find movie based on folder names and MovieQueue table
         '''
 
         for file in files['files']:
@@ -214,11 +294,14 @@ class RenamerCron(cronBase):
             dirnames.reverse()
 
             for dir in dirnames:
-                # check and see if name is in history
-                history = Db.query(History).filter_by(name = dir).first()
-                if history:
-                    log.info('Found movie via History.')
-                    return history.movie
+                # check and see if name is in queue
+                queue = Db.query(MovieQueue).filter_by(name = dir).first()
+                if queue:
+                    log.info('Found movie via MovieQueue.')
+                    return {
+                        'queue': queue,
+                        'movie': queue.Movie
+                    }
 
                 # last resort, match every word in path to db
                 lastResort = {}
@@ -240,7 +323,10 @@ class RenamerCron(cronBase):
 
                     if wordCount == len(words) and len(words) > 0:
                         log.info('Found via last resort searching.')
-                        return l
+                        return {
+                            'queue': None,
+                            'movie': l
+                        }
             # Try finding movie on theMovieDB
             # more checking here..
 
@@ -266,14 +352,10 @@ class RenamerCron(cronBase):
 
                 subfiles = {'nfo':{}, 'files':[], 'subtitles':[]}
 
-                movieExt = ['*.mkv', '*.wmv', '*.avi', '*.mpg', '*.mpeg', '*.mp4', '*.m2ts', '*.iso']
-                nfoExt = ['*.nfo']
-                subExt = ['*.sub', '*.srt', '*.idx', '*.ssa', '*.ass']
-
                 patterns = []
-                patterns.extend(movieExt)
-                patterns.extend(nfoExt)
-                patterns.extend(subExt)
+                patterns.extend(self.movieExt)
+                patterns.extend(self.nfoExt)
+                patterns.extend(self.subExt)
 
                 for pattern in patterns:
                     for filename in fnmatch.filter(filenames, pattern):
@@ -285,10 +367,10 @@ class RenamerCron(cronBase):
                         }
 
                         #nfo file
-                        if('*.' + new.get('ext') in nfoExt):
+                        if('*.' + new.get('ext') in self.nfoExt):
                             subfiles['nfo'] = new
                         #subtitle file
-                        elif('*.' + new.get('ext') in subExt):
+                        elif('*.' + new.get('ext') in self.subExt):
                             subfiles['subtitles'].append(new)
                         else:
                             #ignore movies files / or not
